+++ src/gui/widgets/docks/properties_dock.cpp
@@ -39,7 +39,7 @@ PropertiesDock::PropertiesDock(GlaxnimateWindow* parent, item_models::PropertyMo
     d->ui.view_properties->header()->setSectionResizeMode(item_models::PropertyModelSingle::ColumnToggleKeyframe, QHeaderView::ResizeToContents);
     d->ui.view_properties->header()->setSectionResizeMode(item_models::PropertyModelSingle::ColumnValue, QHeaderView::Stretch);
 
-    new ValueDragEventFilter(d->ui.view_properties, item_models::PropertyModelSingle::ColumnValue);
+    auto filter = new ValueDragEventFilter(d->ui.view_properties, item_models::PropertyModelSingle::ColumnValue);
 
     connect(this, &QWidget::customContextMenuRequested, this, &PropertiesDock::custom_context_menu);
 }