+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     * Gradient properties are now shown on the timeline
     * Added context menu for more objects
     * Context menu has been re-organized with more sub-menus for ease of navigation
+    * Both timeline and properties dock now show the value editor on a single click
 * I/O:
     * Improved matte import and export for lottie
     * Improved mask properties import for lottie